chore: simplify based on team discussion yesterday
