+++ packages/angular/common/src/index.ts
@@ -1,9 +1,9 @@
 export { AlertController } from './providers/alert-controller';
+export { AnimationController } from './providers/animation-controller';
+export { GestureController } from './providers/gesture-controller';
 export { LoadingController } from './providers/loading-controller';
 export { MenuController } from './providers/menu-controller';
 export { PickerController } from './providers/picker-controller';
-
-export { GestureController } from './providers/gesture-controller';
 export { DomController } from './providers/dom-controller';
 export { NavController } from './providers/nav-controller';
